requirements.txt
python-dotenv>=1.0.1
openai>=1.40.0
pydantic>=2.7.0
requests>=2.32.0
